pyproject.toml
[project]
name = "repl-nix-newssummarizer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0.11",
    "newspaper3k>=0.2.8",
    "openai>=1.53.1",
    "pandas>=2.2.3",
    "pytz>=2024.2",
    "trafilatura>=1.12.2",
]
